+++ src/rs_driver/api/lidar_driver.h
@@ -27,7 +27,7 @@ namespace robosense
 namespace lidar
 {
 /**
- * @description:  This is the RoboSense LiDAR driver interface class.
+ * @brief:  This is the RoboSense LiDAR driver interface class.
  *                 Now support RS16,RS32, RSBP & RS128.
  */
 template <typename PointT>
@@ -35,103 +35,112 @@ class LidarDriver
 {
 public:
   /**
-   * @description: Constructor, instanciate the driver pointer.
+   * @brief: Constructor, instanciate the driver pointer.
    */
   LidarDriver()
   {
     driver_ptr_ = std::make_shared<LidarDriverImpl<PointT>>();
-  };
+  }
+
   /**
-   * @description: Deconstructor, stop all threads
+   * @brief: Deconstructor, stop all threads
    */
   ~LidarDriver()
   {
     stop();
   }
+
   /**
-   * @description: The initialize function, used to set the realated parameters and instance objects,
-   *               used when get packets from online lidar or pcap.
-   * @param The struct->RSDriverParam
-   * @return: bool
+   * @brief The initialize function, used to set the realated parameters and instance objects,
+   *        used when get packets from online lidar or pcap
+   * @param param The custom struct RSDriverParam
+   * @return If success, return ture; else return false
    */
   inline bool init(const RSDriverParam& param)
   {
     return driver_ptr_->init(param);
   }
+
   /**
-   * @description: The initialize function, only initilize decoder(not include input module),
-   *               only be used when not get packets from ROS or other ways excluding online lidar and pcap.
-   * @param The struct->RSDriverParam
+   * @brief The initialize function which only initialize decoder(not include input module). If lidar packets are from
+   * ROS or other ways excluding online lidar and pcap, call this function to initialize instead of calling init()
+   * @param param The custom struct RSDriverParam
    */
   inline void initDecoderOnly(const RSDriverParam& param)
   {
     driver_ptr_->initDecoderOnly(param);
   }
+
   /**
-   * @description: Start the thread to receive packets, and decode packets
-   * @return: bool
+   * @brief Start the thread to receive packets, and decode packets
+   * @return If success, return ture; else return false
    */
   inline bool start()
   {
     return driver_ptr_->start();
   }
+
   /**
-   * @description: Stop all threads
+   * @brief: Stop all threads
    */
   inline void stop()
   {
     driver_ptr_->stop();
   }
+
   /**
-   * @description: Register the lidar point cloud callback function.
-   *  When pointcloud is prepared, this function will be called.
-   * @param callBack the callback funtion
+   * @brief Register the lidar point cloud callback function to driver. When point cloud is ready, this function will be
+   * called
+   * @param callback The callback funtion
    */
-  inline void regRecvCallback(const std::function<void(const PointcloudMsg<PointT>&)> callBack)
+  inline void regRecvCallback(const std::function<void(const PointCloudMsg<PointT>&)> callback)
   {
-    driver_ptr_->regRecvCallback(callBack);
+    driver_ptr_->regRecvCallback(callback);
   }
+
   /**
-   * @description: Register the lidar scan message callback funtion.
-   * When lidar scan message is ready, this function will be called.
-   * @param callBack the callback funtion
+   * @brief Register the lidar scan message callback funtion to driver.When lidar scan message is ready, this function
+   * will be called
+   * @param callback The callback funtion
    */
-  inline void regRecvCallback(const std::function<void(const ScanMsg&)> callBack)
+  inline void regRecvCallback(const std::function<void(const ScanMsg&)> callback)
   {
-    driver_ptr_->regRecvCallback(callBack);
+    driver_ptr_->regRecvCallback(callback);
   }
+
   /**
-   * @description: Register the lidar difop packet message callback funtion.
-   * When lidar difop packet message is ready, this function will be called.
-   * @param callBack the callback funtion
+   * @brief Register the lidar difop packet message callback funtion to driver. When lidar difop packet message is
+   * ready, this function will be called
+   * @param callback The callback funtion
    */
-  inline void regRecvCallback(const std::function<void(const PacketMsg&)> callBack)
+  inline void regRecvCallback(const std::function<void(const PacketMsg&)> callback)
   {
-    driver_ptr_->regRecvCallback(callBack);
+    driver_ptr_->regRecvCallback(callback);
   }
+
   /**
-   * @description: Register the exception message callback funtion.
-   * When error occurs, this function will be called.
-   * @param excallBack The callback funtion
+   * @brief Register the exception message callback funtion to driver. When error occurs, this function will be called
+   * @param callback The callback funtion
    */
-  inline void regExceptionCallback(const std::function<void(const Error&)> excallBack)
+  inline void regExceptionCallback(const std::function<void(const Error&)> callback)
   {
-    driver_ptr_->regExceptionCallback(excallBack);
+    driver_ptr_->regExceptionCallback(callback);
   }
+
   /**
-   * @description: Decode the scan message. Can be called when processing offline lidar message.
-   **NOTE** This function will only work after decodeDifopPkt is called,
-            because the driver need difop packet to help to decode scan message.
-   * @param pkt_scan_msg The lidar scan message used to be decode
-   * @param point_msg The output point cloud message
-   * @return: bool
+   * @brief Decode the lidar scan message to point cloud
+   * @note This function will only work after decodeDifopPkt is called unless wait_for_difop is set to false
+   * @param pkt_scan_msg The lidar scan message
+   * @param point_cloud_msg The output point cloud message
+   * @return if decode success, return true; else return false
    */
-  inline bool decodeMsopScan(const ScanMsg& pkt_scan_msg, PointcloudMsg<PointT>& point_msg)
+  inline bool decodeMsopScan(const ScanMsg& pkt_scan_msg, PointCloudMsg<PointT>& point_msg)
   {
     return driver_ptr_->decodeMsopScan(pkt_scan_msg, point_msg);
   }
+
   /**
-   * @description: Decode the lidar difop message. **Must** be called when processing offline lidar message.
+   * @brief: Decode the lidar difop message
    * @param pkt_msg The lidar difop packet
    */
   inline void decodeDifopPkt(const PacketMsg& pkt_msg)